+++ test/data/test_multi_nested_tensor.py
@@ -1,5 +1,5 @@
 import random
-from typing import List
+from typing import List, Union
 
 import pytest
 import torch
@@ -18,6 +18,19 @@ def assert_equal(tensor_mat: List[List[Tensor]],
             assert torch.allclose(tensor_mat[i][j], tensor)
 
 
+def column_select(
+    tensor_mat: List[List[Tensor]],
+    index: Union[List[int], slice],
+) -> List[List[Tensor]]:
+    new_tensor_mat = []
+    for tensor_vec in tensor_mat:
+        if isinstance(index, slice):
+            new_tensor_mat.append(tensor_vec[index])
+        else:
+            new_tensor_mat.append([tensor_vec[idx] for idx in index])
+    return new_tensor_mat
+
+
 def test_multi_nested_tensor_basic():
     num_rows = 8
     num_cols = 10
@@ -31,8 +44,8 @@ def test_multi_nested_tensor_basic():
         tensor_mat.append(tensor_list)
 
     multi_nested_tensor = MultiNestedTensor.from_tensor_mat(tensor_mat)
-    assert str(multi_nested_tensor
-               ) == "MultiNestedTensor(num_rows=8, num_cols=10, device='cpu')"
+    assert (str(multi_nested_tensor) ==
+            "MultiNestedTensor(num_rows=8, num_cols=10, device='cpu')")
 
     # Test sizes
     assert multi_nested_tensor.shape[0] == num_rows
@@ -78,7 +91,7 @@ def test_multi_nested_tensor_basic():
                  multi_nested_tensor.narrow(dim=0, start=3, length=2))
 
     # Test multi_nested_tensor[List[int]] indexing
-    for index in [[4], [2, 2], [-4, 1, 7], [3, -7, 1, 0]]:
+    for index in [[4], [2, 2], [-4, 1, 7], [3, -7, 1, 0], []]:
         multi_nested_tensor_indexed = multi_nested_tensor[index]
         assert multi_nested_tensor_indexed.shape[0] == len(index)
         assert multi_nested_tensor_indexed.shape[1] == num_cols
@@ -92,25 +105,85 @@ def test_multi_nested_tensor_basic():
         multi_nested_tensor_col = multi_nested_tensor[:, j]
         assert multi_nested_tensor_col.shape[0] == num_rows
         assert multi_nested_tensor_col.shape[1] == 1
-        for i in range(-num_rows, num_rows):
-            tensor = multi_nested_tensor_col[i, 0]
-            assert isinstance(tensor, torch.Tensor)
-            assert torch.allclose(tensor_mat[i][j], tensor)
+        assert_equal(column_select(tensor_mat, [j]), multi_nested_tensor_col)
 
     # Test column List[int] indexing
-    with pytest.raises(NotImplementedError):
-        # TODO: Add proper test once implemented
-        multi_nested_tensor[:, [2, 4, 6]]
+    for index in [[4], [2, 2], [-4, 1, 7], [3, -7, 1, 0], []]:
+        assert_equal(column_select(tensor_mat, index),
+                     multi_nested_tensor[:, index])
 
     # Test column slicing
-    with pytest.raises(NotImplementedError):
-        # TODO: Add proper test once implemented
-        multi_nested_tensor[:, 4:8]
+    assert_equal(tensor_mat, multi_nested_tensor[:, :])
+    assert_equal(column_select(tensor_mat, slice(None, 3)),
+                 multi_nested_tensor[:, :3])
+    assert_equal(column_select(tensor_mat, slice(3, None)),
+                 multi_nested_tensor[:, 3:])
+    assert_equal(column_select(tensor_mat, slice(3, 5)),
+                 multi_nested_tensor[:, 3:5])
+    assert_equal(column_select(tensor_mat, slice(-7, 5)),
+                 multi_nested_tensor[:, -7:5])
+    assert_equal(column_select(tensor_mat, slice(-7, -1)),
+                 multi_nested_tensor[:, -7:-1])
+    assert_equal(column_select(tensor_mat, slice(1, None, 2)),
+                 multi_nested_tensor[:, 1::2])
+    empty_multi_nested_tensor = multi_nested_tensor[:, 5:3]
+    assert empty_multi_nested_tensor.shape[0] == num_rows
+    assert empty_multi_nested_tensor.shape[1] == 0
 
     # Test column narrow
-    with pytest.raises(NotImplementedError):
-        # TODO: Add proper test once implemented
-        multi_nested_tensor.narrow(dim=1, start=3, length=2)
+    assert_equal(column_select(tensor_mat, slice(3, 3 + 2)),
+                 multi_nested_tensor.narrow(dim=1, start=3, length=2))
+
+    # Testing row concat
+    assert_equal(
+        tensor_mat,
+        MultiNestedTensor.cat(
+            (multi_nested_tensor[:2], multi_nested_tensor[2:4],
+             multi_nested_tensor[4:]), dim=0),
+    )
+    assert_equal(
+        tensor_mat,
+        MultiNestedTensor.cat([
+            multi_nested_tensor[i] for i in range(multi_nested_tensor.size(0))
+        ], dim=0),
+    )
+    assert_equal(tensor_mat, MultiNestedTensor.cat([multi_nested_tensor],
+                                                   dim=0))
+    with pytest.raises(RuntimeError, match="num_cols must be the same"):
+        MultiNestedTensor.cat([
+            multi_nested_tensor[:2],
+            multi_nested_tensor[2:4, 0],
+        ], dim=0)
+    with pytest.raises(RuntimeError, match="Cannot concatenate"):
+        MultiNestedTensor.cat([], dim=0)
+
+    # Testing col concat
+    assert_equal(
+        tensor_mat,
+        MultiNestedTensor.cat(
+            (multi_nested_tensor[:, :2], multi_nested_tensor[:, 2:4],
+             multi_nested_tensor[:, 4:]), dim=1),
+    )
+    assert_equal(
+        tensor_mat,
+        MultiNestedTensor.cat([
+            multi_nested_tensor[:, i]
+            for i in range(multi_nested_tensor.size(1))
+        ], dim=1),
+    )
+    assert_equal(tensor_mat, MultiNestedTensor.cat([multi_nested_tensor],
+                                                   dim=1))
+    with pytest.raises(RuntimeError, match="num_rows must be the same"):
+        MultiNestedTensor.cat([
+            multi_nested_tensor[1:],
+            multi_nested_tensor,
+        ], dim=1)
+    with pytest.raises(RuntimeError, match="Cannot concatenate"):
+        MultiNestedTensor.cat([], dim=1)
+
+    # Testing set item
+    with pytest.raises(RuntimeError, match="read-only"):
+        multi_nested_tensor[0, 0] = torch.zeros(3)
 
     # Testing clone
     cloned_multi_nested_tensor = multi_nested_tensor.clone()