+++ torch_frame/data/stats.py
@@ -30,26 +30,16 @@ class StatType(Enum):
 
     @staticmethod
     def stats_for_stype(stype: torch_frame.stype) -> List['StatType']:
-        if stype == torch_frame.numerical:
-            return [
+        stats_type = {
+            torch_frame.numerical: [
                 StatType.MEAN,
                 StatType.STD,
                 StatType.QUANTILES,
-            ]
-        elif stype == torch_frame.categorical:
-            return [
-                StatType.COUNT,
-            ]
-        elif stype == torch_frame.text_embedded:
-            return []
-        elif stype == torch_frame.text_tokenized:
-            return []
-        elif stype == torch_frame.multicategorical:
-            return [
-                StatType.MULTI_COUNT,
-            ]
-
-        raise NotImplementedError(f"Invalid semantic type '{stype.value}'")
+            ],
+            torch_frame.categorical: [StatType.COUNT],
+            torch_frame.multicategorical: [StatType.MULTI_COUNT]
+        }
+        return stats_type.get(stype, [])
 
     def compute(self, ser: Series, sep: Optional[str] = None) -> Any:
         if self == StatType.MEAN:
@@ -64,16 +54,15 @@ class StatType(Enum):
         elif self == StatType.COUNT:
             count = ser.value_counts(ascending=False)
             return count.index.tolist(), count.values.tolist()
+
         elif self == StatType.MULTI_COUNT:
             assert sep is not None
-            print("sep is ", sep)
             ser = ser.apply(
                 lambda x: set([cat.strip() for cat in x.split(sep)])
                 if (x is not None and x != '') else set())
             ser = ser.explode().dropna()
             count = ser.value_counts(ascending=False)
             return count.index.tolist(), count.values.tolist()
-        raise NotImplementedError(f"Invalid stat type '{self.value}'")
 
 
 def compute_col_stats(