+++ torch_frame/stype.py
@@ -25,6 +25,7 @@ class stype(Enum):
     text_embedded = 'text_embedded'
     text_tokenized = 'text_tokenized'
     multicategorical = 'multicategorical'
+    sequence_numerical = 'sequence_numerical'
 
     @property
     def is_text_stype(self) -> bool:
@@ -35,7 +36,7 @@ class stype(Enum):
         r""" This property indicates if the data of an stype is stored in
         :class:`torch_frame.data.MultiNestedTensor`.
         """
-        return self in [stype.multicategorical]
+        return self in [stype.multicategorical, self.sequence_numerical]
 
     @property
     def use_dict_multi_nested_tensor(self) -> bool:
@@ -50,3 +51,4 @@ categorical = stype('categorical')
 text_embedded = stype('text_embedded')
 text_tokenized = stype('text_tokenized')
 multicategorical = stype('multicategorical')
+sequence_numerical = stype('sequence_numerical')